Publish MarketDataCollectedEvent to ReddisMQ on save. Resolves #13

+++ FinancialPlatform/FinancialPlatform.Contracts/MarketDataCollectedEvent.cs
@@ -1,3 +1,3 @@
 ﻿namespace FinancialPlatform.Contracts;
 
-public record MarketDataCollectedEvent(string Ticker, int Price, DateTime TimeStamp);
+public record MarketDataCollectedEvent(string Ticker, decimal Price, DateTime TimeStamp);
